extracted select input component to add testing
